Palette is now partially working in Donkey Kong
Rendering is now disabled after reset or power on
Increased the amount of time the PPU sleeps during startup.
During startup the PPU will now accurately increment the cycle and scanlines
Disabling rendering now only affects the innercycle. This was causing the Vblank flags to not work properly.
Made a change to the palette rendering to more accurately reflect other emulators.